moved hook into component folder
